Module owlc_passes::resolver

Resolving logic for symbols in the generated abstract syntax tree.

Structs

ResolverVisitor
Generate a map for resolving symbols in the abstract syntax tree.

SymbolTable
Manages all the symbols that are currently visible.

Enums

Scope
Represents a lexical scope. A new Scope should be created every time the set of visible symbols changes. This can happen at blocks and let statements. For performance reasons, Block variant can contain multiple Symbols for function arguments. Example:

Symbol
Represents a resolved symbol (can be either variable type or function type).